+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 
 ## [Unreleased]
+- Added: New "api_gateway_minimum_compression_size" to allow compression configuration on the APIGateway
+- Added: New variables to control ECS scaling adjustments
+	- api_ecs_autoscale_scale_down_adjustment
+	- api_ecs_autoscale_scale_up_adjustment
+	- push_ecs_autoscale_scale_down_adjustment
+	- push_ecs_autoscale_scale_up_adjustment
 
 
 ## [v0.1.7] 2020-09-17
@@ -16,7 +22,7 @@ All notable changes to this project will be documented in this file.
 - Added: Added "db_pool_size" variable to control pg max pool size - Can be used by ECS API and ECS Push services
 - Updated: Allow stats lambda to access "time_zone" parameter
 - Updated: Added additional resources to the operators group policy for MFA
-- Fix: Adjusted values displayed in the CloudWatch dashboard, so take into consideration doubled logs in ApiGateway logs 
+- Fix: Adjusted values displayed in the CloudWatch dashboard, so take into consideration doubled logs in ApiGateway logs
 
 
 ## [v0.1.6] 2020-09-10